Add license and fix bugs (#112)
* Update README with direnv (#105)

* Update urllib minor version (#106)

* Add MiT License (#107)

* Add export to variables for direnv (#109)

* Fix swipe data parsing error (#110)

Encountered a bug where data.log (stores the queries from cornell dining
requests every 5 minutes) contains two logs of {"msg": "Unauthorized"}
which created a key error when we tried to access 'TIMESTAMP' from the
log. This exception was not caught and prevented further updating of
our csv files. This bug may also be responsible for inconsistent ctown
data. We fix the issue by checking the 'TIMESTAMP' is a valid key in the
log.

* Add new meal plans and check for unlimited (#111)