Fix dates for swipe data and error checking for Get requests (#139)
* Update README with direnv (#105)

* Update urllib minor version (#106)

* Add MiT License (#107)

* Add export to variables for direnv (#109)

* Fix swipe data parsing error (#110)

Encountered a bug where data.log (stores the queries from cornell dining
requests every 5 minutes) contains two logs of {"msg": "Unauthorized"}
which created a key error when we tried to access 'TIMESTAMP' from the
log. This exception was not caught and prevented further updating of
our csv files. This bug may also be responsible for inconsistent ctown
data. We fix the issue by checking the 'TIMESTAMP' is a valid key in the
log.

* Add new meal plans and check for unlimited (#111)

* Update vulnerability (#114)

* Add sources and new URL (#121)

* Change URLs for Static JSON Files (#122)

* Add padding to BRB values (#123)

* Check if transactions are positive (#124)

* Allow string inputs for Get swipes (#125)

* Change 'if (x)' to 'if x' (#126)

* Preload data before starting gunicorn workers (#129)

* Add Black formatter (#127)

* Clean import statements & docstrings (#128)

* Make swipe dates ISO 8601 (#130)

* Skip lines in swipe logs that aren't formatted correctly (#136)

* Add error checking to Get requests (#137)

* Change date in export_data (#138)